# src/Prdb.Ordeno.Host/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "prdb-ordeno",
    "description": "The API the browser side of prdb-ordeno talks to. Generated from the code at build time and committed: change an endpoint, not this file.",
    "version": "v1"
  },
  "paths": {
...
    "/api/configuration/api-key": {
      "put": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetApiKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/sources": {
      "post": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/sources/{id}": {
      "delete": {
        "tags": [
          "Configuration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/target": {
      "put": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/artwork": {
      "put": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetArtworkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/unattended-filing": {
      "put": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUnattendedFilingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/media-server": {
      "put": {
        "tags": [
          "Configuration"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMediaServerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaServerCheckState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Configuration"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/media-server/test": {
      "post": {
        "tags": [
          "Configuration"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaServerCheckState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
    "/api/configuration/completion": {
      "post": {
        "tags": [
          "Configuration"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationState"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationProblem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddSourceRequest": {
        "required": [
          "path"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
...
      "ConfigurationProblem": {
        "required": [
          "message",
          "configuration"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "configuration": {
            "$ref": "#/components/schemas/ConfigurationState"
          }
        }
      },
      "ConfigurationState": {
        "required": [
          "apiKeySet",
          "sources",
          "target",
          "layout",
          "availableLayouts",
          "mediaServer",
          "artwork",
          "unattended",
          "unattendedIntervalMinutes",
          "complete",
          "readyToComplete",
          "whatHappensNext"
        ],
        "type": "object",
        "properties": {
          "apiKeySet": {
            "type": "boolean"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceState"
            }
          },
          "target": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DirectoryState"
              }
            ]
          },
          "layout": {
            "type": [
              "null",
              "string"
            ]
          },
          "availableLayouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LayoutOption"
            }
          },
          "mediaServer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MediaServerState"
              }
            ]
          },
          "artwork": {
            "type": "boolean"
          },
          "unattended": {
            "type": "boolean"
          },
          "unattendedIntervalMinutes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "complete": {
            "type": "boolean"
          },
          "readyToComplete": {
            "type": "boolean"
          },
          "whatHappensNext": {
            "type": "string"
          }
        }
      },
      "DirectoryState": {
        "required": [
          "path",
          "usable",
          "problem"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "usable": {
            "type": "boolean"
          },
          "problem": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LayoutOption": {
        "required": [
          "name",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "MediaServerCheckState": {
        "required": [
          "status",
          "message",
          "working",
          "configuration"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "working": {
            "type": "boolean"
          },
          "configuration": {
            "$ref": "#/components/schemas/ConfigurationState"
          }
        }
      },
      "MediaServerState": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
...
      "SetApiKeyRequest": {
        "required": [
          "apiKey"
        ],
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string"
          }
        }
      },
      "SetArtworkRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "SetMediaServerRequest": {
        "required": [
          "url",
          "apiKey"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          }
        }
      },
...
      "SetTargetRequest": {
        "required": [
          "path",
          "layout"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "layout": {
            "type": "string"
          }
        }
      },
      "SetUnattendedFilingRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        }
      },
...
      "SourceState": {
        "required": [
          "id",
          "path",
          "usable",
          "problem",
          "movement",
          "movementExplained"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "path": {
            "type": "string"
          },
          "usable": {
            "type": "boolean"
          },
          "problem": {
            "type": [
              "null",
              "string"
            ]
          },
          "movement": {
            "type": "string"
          },
          "movementExplained": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}